Use ssize_t, the same as filesize, which will always be less as block numbers

+++ dvd_cell.c
@@ -44,7 +44,7 @@ uint32_t dvd_cell_last_sector(const ifo_handle_t *vmg_ifo, const ifo_handle_t *v
 
 }
 
-uint32_t dvd_cell_blocks(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number) {
+ssize_t dvd_cell_blocks(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number) {
 
 	uint32_t first_sector = 0;
 	first_sector = dvd_cell_first_sector(vmg_ifo, vts_ifo, track_number, cell_number);
@@ -52,7 +52,7 @@ uint32_t dvd_cell_blocks(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_if
 	uint32_t last_sector = 0;
 	last_sector = dvd_cell_last_sector(vmg_ifo, vts_ifo, track_number, cell_number);
 
-	uint32_t blocks = 0;
+	ssize_t blocks = 0;
 	blocks = last_sector - first_sector;
 
 	// Include the last cell
@@ -65,7 +65,7 @@ uint32_t dvd_cell_blocks(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_if
 
 ssize_t dvd_cell_filesize(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo, const uint16_t track_number, const uint8_t cell_number) {
 
-	uint32_t blocks;
+	ssize_t blocks;
 	blocks = dvd_cell_blocks(vmg_ifo, vts_ifo, track_number, cell_number);
 
 	ssize_t filesize = 0;
